fix: 스킬 평균 계산의 BigDecimal 정밀도 보존 (#236)
- AVG projection 반환 타입을 Double에서 BigDecimal로 변경
- 스킬 평균 계산 과정의 부동소수점 변환 제거
- HALF_UP 반올림 경계값 및 PostgreSQL AVG 매핑 테스트 추가

+++ src/main/java/com/mr/domain/analysis/repository/AnalysisRepository.java
@@ -4,6 +4,7 @@ import com.mr.domain.analysis.entity.Analysis;
 import com.mr.domain.analysis.entity.enums.AnalysisStatus;
 import jakarta.persistence.LockModeType;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
@@ -122,10 +123,10 @@ public interface AnalysisRepository extends JpaRepository<Analysis, Long> {
     }
 
     interface WeeklySkillAverages {
-        Double getScaleScore();
-        Double getTensionScore();
-        Double getProgressionScore();
-        Double getVoiceLeadingScore();
+        BigDecimal getScaleScore();
+        BigDecimal getTensionScore();
+        BigDecimal getProgressionScore();
+        BigDecimal getVoiceLeadingScore();
     }
 
     @Modifying(clearAutomatically = true)
